feat: remove providers page and move it to settings
